requirements.txt
glfw==2.7.0
numpy==2.2.0
Pillow==11.0.0
PyOpenGL==3.1.7
warp_lang==1.4.2